requirements.txt
pyrogram
TgCrypto
yt-dlp
requests
Pillow